V0.1.6 Fixed crash with receiving items of extreme value

+++ Source/APCpp/inc/Archipelago.h
@@ -243,7 +243,7 @@ struct AP_Gift {
     std::string ID;
     std::string ItemName;
     size_t Amount;
-    size_t ItemValue;
+    int64_t ItemValue;
     std::vector<AP_GiftTrait> Traits;
     std::string Sender;
     std::string Receiver;
